travel-lykkr-github/requirements.txt
Django==5.2.1
mysqlclient==2.2.4
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
